+++ src/main/java/dev/gustavo/math/repository/SubmissionRepository.java
@@ -3,5 +3,5 @@ package dev.gustavo.math.repository;
 import dev.gustavo.math.entity.Submission;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubmissionRepository extends JpaRepository<Submission, Integer> {
+public interface SubmissionRepository extends JpaRepository<Submission, Long> {
 }
