+++ src/main/java/dev/gustavo/math/repository/TestCaseRepository.java
@@ -3,5 +3,5 @@ package dev.gustavo.math.repository;
 import dev.gustavo.math.entity.TestCase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TestCaseRepository extends JpaRepository<TestCase, Integer> {
+public interface TestCaseRepository extends JpaRepository<TestCase, Long> {
 }
